requirements.txt
Werkzeug
Flask
numpy
Keras
gevent
pillow
h5py
tensorflow-cpu>=2.1